Chatbot/requirements.txt
python-dotenv==1.0.0
streamlit==1.32.0
faiss-cpu==1.7.4
sentence-transformers==2.5.1
numpy==1.26.4
pandas==2.2.1 
flask==2.3.3
flask-cors==4.0.0
google-generativeai
pymongo==4.6.3
